Old files/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pycryptodome==3.23.0
Pillow>=9.0.0
